Better flow propagation with increased window size

+++ trackers/base_klt.py
@@ -127,7 +127,7 @@ class BaseKLT(object):
         for col, pts in izip(cols.astype(np.int64), self.tm_.tracks.itervalues()): 
             cv2.polylines(out, [np.vstack(pts.items).astype(np.int32)[-max_track_length:]], False, 
                           tuple(col), thickness=1)
-            tl, br = pts.latest_item-2, pts.latest_item+2
+            tl, br = np.int32(pts.latest_item)-2, np.int32(pts.latest_item)+2
             cv2.rectangle(out, (tl[0], tl[1]), (br[0], br[1]), tuple(col), -1)
 
     def viz(self, out, colored=False):
